botproduto/requirements.txt
asttokens==2.4.1
botcity-email-plugin==0.1.5
botcity-http-plugin==0.1.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
idna==3.10
imap-tools==1.7.2
ipython==8.27.0
jedi==0.19.1
matplotlib-inline==0.1.7
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
reportlab==4.2.2
requests==2.32.3
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webdriver-manager==4.0.2
